feature/Profile Screen Adjustments
