+++ HandleService/app/src/main/java/br/com/handleservice/presentation/screens/profile/ProfileScreen.kt
@@ -2,7 +2,6 @@ package br.com.handleservice.presentation.screens.profile
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.*
@@ -19,6 +18,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import br.com.handleservice.R
+import br.com.handleservice.presentation.screens.profile.components.ProfileOption
 
 @Composable
 fun ProfileScreen(
@@ -30,7 +30,7 @@ fun ProfileScreen(
         modifier = Modifier
             .fillMaxSize()
             .background(colorResource(R.color.background))
-            .padding(horizontal = 16.dp),
+            .padding(horizontal = 30.dp),
     ) {
         Spacer(modifier = Modifier.height(24.dp))
 
@@ -51,20 +51,14 @@ fun ProfileScreen(
             Spacer(modifier = Modifier.width(16.dp))
             Text(
                 text = profileState.name,
-                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.Bold),
-                fontSize = 20.sp,
-                color = colorResource(R.color.handle_titles)
+                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.W600),
+                fontSize = 21.sp,
+                color = colorResource(R.color.black)
             )
         }
 
         Spacer(modifier = Modifier.height(12.dp))
 
-        ProfileOption(
-            icon = R.drawable.ic_filled_messages,
-            label = "Chat",
-            description = "Meu histórico de conversas",
-            onClick = { }
-        )
         ProfileOption(
             icon = R.drawable.ic_filled_address,
             label = "Meus Endereços",
@@ -72,8 +66,14 @@ fun ProfileScreen(
             onClick = { }
         )
         ProfileOption(
-            icon = R.drawable.ic_filled_document_data,
-            label = "Dados da Conta",
+            icon = R.drawable.ic_favorite,
+            label = "Favoritos",
+            description = "Os seus preferidos estarão aqui!",
+            onClick = { }
+        )
+        ProfileOption(
+            icon = R.drawable.ic_notification,
+            label = "Notificações",
             description = "Minhas informações",
             onClick = { }
         )
@@ -82,71 +82,6 @@ fun ProfileScreen(
             label = "Configurações",
             description = "Faça ajustes no seu app",
             onClick = { }
-        )
-    }
-}
-
-@Composable
-fun ProfileOption(
-    icon: Int,
-    label: String,
-    description: String,
-    onClick: () -> Unit
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(onClick = onClick)
-                .padding(horizontal = 16.dp, vertical = 8.dp)
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 16.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Icon(
-                    painter = painterResource(id = icon),
-                    contentDescription = label,
-                    modifier = Modifier.size(36.dp),
-                    tint = colorResource(R.color.handle_light_gray)
-                )
-                Spacer(modifier = Modifier.width(16.dp))
-                Column(
-                    modifier = Modifier.weight(1f)
-                ) {
-                    Text(
-                        text = label,
-                        style = MaterialTheme.typography.bodyMedium.copy(fontWeight = FontWeight.Normal),
-                        fontSize = 18.sp,
-                        color = colorResource(R.color.handle_titles)
-                    )
-                    Text(
-                        text = description,
-                        style = MaterialTheme.typography.bodySmall,
-                        fontSize = 14.sp,
-                        color = colorResource(R.color.handle_gray)
-                    )
-                }
-                Icon(
-                    painter = painterResource(R.drawable.ic_chevron_right),
-                    contentDescription = "Seta",
-                    modifier = Modifier.size(20.dp),
-                    tint = colorResource(R.color.handle_gray)
-                )
-            }
-        }
-        Divider(
-            color = colorResource(R.color.handle_divider),
-            thickness = 1.dp,
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.BottomCenter)
         )
     }
 }
